// src/pages/bookv2/hooks/useBookingPricing.generated.ts
export enum DiscountType {
  Lmd = 'LMD',
  /** LAST MINUTE DEAL */
  None = 'NONE'
}
// ...
export type BookingPricingQuery = { bookingPricing: { commission: number, currency: string, discount?: number | null, discountType?: Types.DiscountType | null, id?: number | null, recommendedRetailPrice: number, totalOrderPrice: number, fareTypesToBook: Array<{ numberOfSeats: number, resSystemFareTypeId: string }>, promocode?: { name: string, discountAmount: number } | null } };